+++ MoMoney.Core/ViewModels/AddTransactionViewModel.cs
@@ -1,6 +1,8 @@
 ﻿using System.Collections.ObjectModel;
 using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.ComponentModel;
+using Laerdal.FFmpeg;
+using Plugin.Maui.Audio;
 using MoMoney.Core.Models;
 using MoMoney.Core.Helpers;
 using MoMoney.Core.Exceptions;
@@ -14,6 +16,11 @@ public partial class AddTransactionViewModel : ObservableObject
     readonly ICategoryService categoryService;
     readonly ITransactionService transactionService;
     readonly ILoggerService<AddTransactionViewModel> logger;
+
+    readonly IAudioManager audioManager;
+    readonly IAudioRecorder recorder;
+
+    readonly IOpenAIService openAIService;
 
     [ObservableProperty]
     public ObservableCollection<Account> accounts = new();
@@ -43,15 +50,27 @@ public partial class AddTransactionViewModel : ObservableObject
     public Category subcategory = new();
 
     [ObservableProperty]
+    public string payee;
+
+    [ObservableProperty]
     public Account transferAccount = new();
 
-    public AddTransactionViewModel(ITransactionService _transactionService, IAccountService _accountService,
-        ICategoryService _categoryService, ILoggerService<AddTransactionViewModel> _logger)
+    TransactionType transactionType = TransactionType.None;
+
+    ResponseIDs responseIDs = null;
+
+    public AddTransactionViewModel(ITransactionService _transactionService, IAccountService _accountService, ICategoryService _categoryService,
+        ILoggerService<AddTransactionViewModel> _logger, IOpenAIService _openAIService, IAudioManager _audioManager)
     {
         transactionService = _transactionService;
         accountService = _accountService;
         categoryService = _categoryService;
         logger = _logger;
+
+        openAIService = _openAIService;
+
+        audioManager = _audioManager;
+        recorder = audioManager.CreateRecorder();
     }
 
     /// <summary>
@@ -85,6 +104,8 @@ public partial class AddTransactionViewModel : ObservableObject
             // re-add selected Subcategory if not null
             if (subcategory is not null)
                 Subcategory = subcategory;
+
+            transactionType = TransactionType.Income;
         }
         catch (CategoryNotFoundException ex)
         {
@@ -111,6 +132,8 @@ public partial class AddTransactionViewModel : ObservableObject
             Subcategories.Clear();
             Categories.Add(transfer);
             Category = transfer;
+
+            transactionType = TransactionType.Transfer;
         }
         catch (CategoryNotFoundException ex)
         {
@@ -147,6 +170,8 @@ public partial class AddTransactionViewModel : ObservableObject
                 Category = savedCategory;
             if (savedSubcategory is not null)
                 Subcategory = savedSubcategory;
+
+            transactionType = TransactionType.Expense;
         }
         catch (Exception ex)
         {
@@ -163,8 +188,7 @@ public partial class AddTransactionViewModel : ObservableObject
     {
         try
         {
-            if (parentCategory is null)
-                return;
+            if (parentCategory is null) return;
 
             var subcategories = await categoryService.GetSubcategories(parentCategory);
             Subcategories.Clear();
@@ -203,26 +227,126 @@ public partial class AddTransactionViewModel : ObservableObject
             Subcategory = Subcategories.First();
     }
 
-    /// <summary>
-    /// adds Category to database using input fields from view.
-    /// </summary>
-    /// <param name="payee"></param>
     [RelayCommand]
-    async Task AddTransaction(string payee)
+    async Task Record(Button btnRecord)
     {
         try
         {
-            if (Account is null || Category is null || Subcategory is null || 
+            if (transactionType == TransactionType.None)
+            {
+                await Utilities.DisplayToast("Please select a transaction type first.");
+                return;
+            }
+
+            string filePath = $"{FileSystem.Current.CacheDirectory}/{Constants.AUDIO_FILE_NAME}";
+            if (recorder.IsRecording)
+            {
+                // stop and check if audio is valid
+                var audioSource = await recorder.StopAsync();
+                
+                using var stream = audioSource.GetAudioStream();
+                var audioPlayer = AudioManager.Current.CreatePlayer(stream);
+                if (audioPlayer == null || audioPlayer.Duration < 3)
+                    throw new FileNotFoundException("Recording is not long enough");
+
+                FFmpegConfig.IgnoreSignal(24); // stops app from freezing
+                FFmpeg.Execute($"-y -i {filePath.Replace("mp3","wav")} -b:a 32k {filePath}");
+
+                var bytes = File.ReadAllBytes(filePath);
+                var audioData = new BinaryData(bytes);
+
+                var transactionResponse = await openAIService.DictateTransaction(audioData, transactionType);
+                if (transactionResponse == null) return;
+
+                responseIDs = transactionResponse.ResponseIDs;
+
+                Date = transactionResponse.Date;
+
+                Amount = transactionResponse.Amount;
+
+                var account = Accounts.FirstOrDefault(a => a.AccountName == transactionResponse.Account);
+                if (account is not null)
+                    Account = account;
+
+                var category = Categories.FirstOrDefault(c => c.CategoryName == transactionResponse.Category);
+                if (category is not null)
+                    Category = category;
+
+                var subcategory = Subcategories.FirstOrDefault(c => c.CategoryName == transactionResponse.Subcategory);
+                if (subcategory is not null)
+                    Subcategory = subcategory;
+
+                Payee = transactionResponse.Payee;
+
+                if (transactionType == TransactionType.Transfer)
+                {
+                    // if contains transfer account, get the account and set transferId
+                    var transferAccount = Accounts.FirstOrDefault(a => a.AccountName == transactionResponse.TransferAccount);
+                    TransferAccount = transferAccount;
+                }
+
+                // delete audio files in cache
+                File.Delete(filePath);
+                File.Delete(filePath.Replace("mp3", "wav"));
+                string[] filesToDelete = Directory.GetFiles(FileSystem.Current.CacheDirectory, "*.tmp");
+                foreach (var file in filesToDelete)
+                    File.Delete(file);
+
+                ResetButtonColour(btnRecord);
+            }
+            else
+            {
+                await recorder.StartAsync(filePath.Replace("mp3", "wav"));
+
+                btnRecord.TextColor = Colors.Red;
+                btnRecord.BorderColor = Colors.Red;
+                btnRecord.Animate("pulse", (d) => btnRecord.FontSize = d, 32, 28, 16, 1000, Easing.Linear, (d, b) => btnRecord.FontSize = 30, () => recorder.IsRecording);
+            }
+        }
+        catch (FileNotFoundException ex)
+        {
+            ResetButtonColour(btnRecord);
+            await logger.LogError(nameof(Record), ex);
+            await Shell.Current.DisplayAlert("Error", "Failed to record audio. Please try again", "OK");
+        }
+        catch (Exception ex)
+        {
+            ResetButtonColour(btnRecord);
+            await logger.LogError(nameof(Record), ex);
+            await Shell.Current.DisplayAlert("Error", ex.Message, "OK");
+        }
+    }
+
+    /// <summary>
+    /// Resets record button colour to default.
+    /// </summary>
+    /// <param name="button"></param>
+    void ResetButtonColour(Button button)
+    {
+        button.SetAppTheme(Button.TextColorProperty, Colors.Black, Colors.White);
+        button.SetAppTheme(Button.BorderColorProperty, (Color)Utilities.Colours["Gray700"], (Color)Utilities.Colours["Gray200"]);
+    }
+
+    /// <summary>
+    /// adds Category to database using input fields from view.
+    /// </summary>
+    [RelayCommand]
+    async Task AddTransaction()
+    {
+        try
+        {
+            if (Account is null || Category is null || Subcategory is null ||
                 (Category.CategoryID == Constants.TRANSFER_ID && TransferAccount is null))
                 return;
 
             // add payee to Payees if not already in list
-            if (!Payees.Contains(payee))
-                Payees.Add(payee);
+            if (!Payees.Contains(Payee))
+                Payees.Add(Payee);
 
+            int ID = -1;
             if (Category.CategoryID == Constants.INCOME_ID) // income = regular
             {
-                await transactionService.AddTransaction(Date, Account.AccountID, Amount,Category.CategoryID, Subcategory.CategoryID, payee, null);
+                ID = await transactionService.AddTransaction(Date, Account.AccountID, Amount, Category.CategoryID, Subcategory.CategoryID, Payee, null);
             }
             else if (Category.CategoryID == Constants.TRANSFER_ID) // transfer = 2 transactions
             {
@@ -232,12 +356,18 @@ public partial class AddTransactionViewModel : ObservableObject
                 var _amount = Amount;
                 var _categoryID = Category.CategoryID;
                 var _transferID = TransferAccount.AccountID;
-                await transactionService.AddTransaction(_date, _accountID, -_amount, _categoryID, Constants.DEBIT_ID, string.Empty, _transferID);
+                ID = await transactionService.AddTransaction(_date, _accountID, -_amount, _categoryID, Constants.DEBIT_ID, string.Empty, _transferID);
                 await transactionService.AddTransaction(_date, _transferID, _amount, _categoryID, Constants.CREDIT_ID, string.Empty, _accountID);
             }
             else if (Category.CategoryID >= Constants.EXPENSE_ID) // expense = negative amount
             {
-                await transactionService.AddTransaction(Date, Account.AccountID, -Amount, Category.CategoryID, Subcategory.CategoryID, payee, null);
+                ID = await transactionService.AddTransaction(Date, Account.AccountID, -Amount, Category.CategoryID, Subcategory.CategoryID, Payee, null);
+            }
+
+            if (ID != -1 && responseIDs != null)
+            {
+                await openAIService.MapDictationToTransaction(ID, responseIDs);
+                responseIDs = null;
             }
 
             ClearAfterAdd();
@@ -270,8 +400,14 @@ public partial class AddTransactionViewModel : ObservableObject
         Category = null;
         Subcategory = null;
         TransferAccount = null;
+        transactionType = TransactionType.None;
 
         Categories.Clear();
         Subcategories.Clear();
+
+        responseIDs = null;
+
+        if (recorder.IsRecording)
+            recorder.StopAsync();
     }
 }